+++ src/blockbenchTypeMods.d.ts
@@ -1,8 +1,8 @@
-import { BlueprintSettings } from './blueprintSettings'
 import type {
 	IBlueprintBoneConfigJSON,
 	IBlueprintLocatorConfigJSON,
 } from './formats/blueprint/format'
+import { BlueprintSettings } from './formats/blueprint/settings'
 import { TextDisplay } from './outliner/textDisplay'
 import { VanillaBlockDisplay } from './outliner/vanillaBlockDisplay'
 import { VanillaItemDisplay } from './outliner/vanillaItemDisplay'
