+++ src/formats/blueprint/dfu.ts
@@ -1,10 +1,10 @@
 import { NbtCompound, NbtList, NbtString, NbtTag } from 'deepslate/lib/nbt'
-import TransparentTexture from '../assets/transparent.png'
-import { PACKAGE } from '../constants'
-import { type IBlueprintFormatJSON, getDefaultProjectSettings } from '../formats/blueprint/format'
-import { openUnexpectedErrorDialog } from '../interface/dialog/unexpectedError'
-import { BoneConfig } from '../nodeConfigs'
-import { MinecraftVersion } from './global'
+import TransparentTexture from '../../assets/transparent.png'
+import { PACKAGE } from '../../constants'
+import { openUnexpectedErrorDialog } from '../../interface/dialog/unexpectedError'
+import { BoneConfig } from '../../nodeConfigs'
+import { MinecraftVersion } from '../../systems/global'
+import { type IBlueprintFormatJSON, getDefaultProjectSettings } from './format'
 
 export function process(model: any): IBlueprintFormatJSON {
 	if (model?.meta?.model_format === 'animatedJava/ajmodel') {
