+++ src/index.ts
@@ -20,9 +20,9 @@ import './systems/minecraft/fontManager'
 import './systems/minecraft/registryManager'
 import './systems/minecraft/versionManager'
 // Misc imports
-import { blueprintSettingErrors } from './blueprintSettings'
 import { BLUEPRINT_CODEC } from './formats/blueprint/codec'
 import { BLUEPRINT_FORMAT } from './formats/blueprint/format'
+import { blueprintSettingErrors } from './formats/blueprint/settings'
 import { openChangelogDialog } from './interface/changelogDialog'
 import { openExportProgressDialog } from './interface/dialog/exportProgress'
 import { openUnexpectedErrorDialog } from './interface/dialog/unexpectedError'
