+++ src/interface/dialog/blueprintSettings.ts
@@ -1,10 +1,10 @@
 import KofiPopup from 'src/components/kofiPopup.svelte'
 import { updateAllCubeOutlines } from 'src/mods/cubeOutlineMod'
 import type { MinecraftVersion } from 'src/systems/global'
-import { defaultValues, type ExportMode } from '../../blueprintSettings'
 import BlueprintSettingsDialogSvelteComponent from '../../components/blueprintSettingsDialog.svelte'
 import { PACKAGE } from '../../constants'
 import { updateRenderBoxPreview, updateRotationLock } from '../../formats/blueprint/format'
+import { defaultValues, type ExportMode } from '../../formats/blueprint/settings'
 import { sanitizeStorageKey } from '../../util/minecraftUtil'
 import { Valuable } from '../../util/stores'
 import { SvelteDialog } from '../../util/svelteDialog'
