+++ src/interface/importAJModelLoader.ts
@@ -2,7 +2,7 @@ import { registerModelLoader } from 'src/util/moddingTools'
 import { SvelteComponentDev } from 'svelte/internal'
 import ImportAjModelLoaderDialog from '../components/importAJModelLoaderDialog.svelte'
 import { BLUEPRINT_CODEC } from '../formats/blueprint/codec'
-import * as ModelDatFixerUpper from '../systems/modelDataFixerUpper'
+import * as ModelDatFixerUpper from '../formats/blueprint/dfu'
 import { injectSvelteComponent } from '../util/injectSvelteComponent'
 import { sanitizeStorageKey } from '../util/minecraftUtil'
 import { translate } from '../util/translation'
