+++ src/systems/exporter.ts
@@ -1,5 +1,5 @@
-import { blueprintSettingErrors } from '../blueprintSettings'
 import { projectTargetVersionIsAtLeast, saveBlueprint } from '../formats/blueprint/format'
+import { blueprintSettingErrors } from '../formats/blueprint/settings'
 import { openBlueprintSettingsDialog } from '../interface/dialog/blueprintSettings'
 import { PROGRESS_DESCRIPTION, openExportProgressDialog } from '../interface/dialog/exportProgress'
 import { openUnexpectedErrorDialog } from '../interface/dialog/unexpectedError'
