+++ src/systems/jsonCompiler.ts
@@ -2,8 +2,8 @@
 /// <reference path="D:/github-repos/snavesutit/blockbench-types/types/index.d.ts"/>
 /// <reference path="../global.d.ts"/>
 
-import { type defaultValues } from '../blueprintSettings'
 import type { IBlueprintBoneConfigJSON } from '../formats/blueprint/format'
+import { type defaultValues } from '../formats/blueprint/settings'
 import {
 	getKeyframeCommands,
 	getKeyframeExecuteCondition,
